+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v3.0.0] - 2023-09-11
+
+### Added
+
+- Added support for healthengine/laravel-prometheus/exporter 4.
+- Added support for Laravel 10.
+- Added support for PHP 8.2.
+
+### Removed
+
+- Removed support for healthengine/laravel-prometheus/exporter 2.
+- Removed support for healthengine/laravel-prometheus/exporter 3.
+- Removed support for Laravel 8.
+- Removed support for Laravel 9.
+- Removed support for PHP 8.0.
 
 ## [v2.0.0] - 2023-02-10
 
@@ -27,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2019-09-20
 
-[unreleased]: https://github.com/HealthEngineAU/laravel-prometheus-collectors/compare/v2.0.0...HEAD
+[v3.0.0]: https://github.com/HealthEngineAU/laravel-prometheus-collectors/compare/v2.0.0...v3.0.0
 [v2.0.0]: https://github.com/HealthEngineAU/laravel-prometheus-collectors/compare/v1.0...v2.0.0
 [v1.0]: https://github.com/HealthEngineAU/laravel-prometheus-collectors/compare/0.4.0...v1.0
 [0.4.0]: https://github.com/HealthEngineAU/laravel-prometheus-collectors/compare/0.3.0...0.4.0